requirements.txt
notebook
ipywidgets
tqdm
python-dotenv
openai==0.27.0
streamlit
aiogram